Ajout Patron de méthode pour l'envoi de données
