+++ app/src/main/java/com/example/florian/altarconquest/View/EcranCreation_Partie.java
@@ -53,10 +53,9 @@ public class EcranCreation_Partie extends Activity implements AdapterView.OnItem
             {
                 Log.i("Création partie", "");
                 //Insertion game dans la bdd
-                ServerSendGameProperties ssgp = new ServerSendGameProperties(EcranCreation_Partie.this);
+                ServerSendGameProperties ssgp = new ServerSendGameProperties();
                 ssgp.execute(addQuote(nomPartie.getText().toString()), addQuote(passwordPartie.getText().toString()), addQuote(String.valueOf(spinner.getSelectedItem())));
-
-
+                
                 ouvrirRejoindrePartie();
             }}
         );
